gst with decimal fields

+++ Appointment/models.py
@@ -223,7 +223,7 @@ class AppointmentCheckout(models.Model):
     rewards =models.ForeignKey(Rewards, on_delete=models.CASCADE, related_name='checkout_reward_appointments', null=True, blank=True) 
     
     tip = models.PositiveIntegerField(default=0, null=True, blank=True)
-    gst = models.PositiveIntegerField(default=0, null=True, blank=True)
+    gst = models.DecimalField(default=0, null=True, blank=True, decimal_places=5, max_digits=5)
     gst_price = models.FloatField(default=0, null=True, blank=True)
     
     service_price = models.PositiveIntegerField(default=0, null=True, blank=True)
